Add Strength of Schedule
Add a strength of schedule component that is calculated by summing the win
count of all FBS opponents and the loss count of all FBS opponents and then
ordering based on (w / w + l).

Also added a flag to each team to indicate whether they are an FBS team. This
flag is set to true if the team shows up in the left hand colum on the input
list.